pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-image-classifier"
version = "1.0.0" 
description = "A flexible, PyTorch-based image classification framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ranjeet Jha", email = "ranjeet.j@gmail.com"},
]
keywords = ["image-classification", "pytorch", "machine-learning", "deep-learning", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "Pillow>=8.0.0",
    "numpy>=1.21.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "scikit-learn>=1.0.0",
    "pandas>=1.3.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.8",
    "mypy>=0.800",
    "pre-commit>=2.15.0",
]
wandb = ["wandb>=0.12.0"]
huggingface = ["huggingface-hub>=0.10.0", "transformers>=4.20.0"]
all = [
    "wandb>=0.12.0",
    "huggingface-hub>=0.10.0",
    "transformers>=4.20.0",
]

[project.urls]
Homepage = "https://github.com/hightechguys79/universal-image-classifier"
Repository = "https://github.com/hightechguys79/universal-image-classifier"
Documentation = "https://github.com/hightechguys79/universal-image-classifier#readme"
"Bug Reports" = "https://github.com/hightechguys79/universal-image-classifier/issues"

[project.scripts]
uic-train = "examples.train_example:main"
uic-infer = "examples.inference_example:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["universal_image_classifier"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
